Implement temporary combo painter with score font, fix time_interval from common_resources

+++ 0.3beta/src/CommonResources.h
@@ -53,7 +53,7 @@ class CommonResources{
   /** Sprites */
   int pieces_width, pieces_height;
 
-  int time_interval;
+  double time_interval;
   int fps;
 
   u_int highscore;
